requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
attrs=22.1.0=pyh71513ae_1
azure-common=1.1.28=pyhd8ed1ab_0
azure-core=1.26.3=pyhd8ed1ab_0
azure-identity=2021.03.15b1=py_0
azure-mgmt-compute=29.1.0=pyh1a96a4e_0
azure-mgmt-core=1.3.2=pyhd8ed1ab_0
azure-mgmt-network=17.0.0=pyhd8ed1ab_0
azure-mgmt-resource=22.0.0=pyhd8ed1ab_0
blinker=1.5=pyhd8ed1ab_0
boto3=1.26.9=pyhd8ed1ab_0
botocore=1.29.10=pyhd8ed1ab_0
brotlipy=0.7.0=py39h02fc5c5_1005
bzip2=1.0.8=h3422bc3_4
ca-certificates=2022.12.7=h4653dfc_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py39h7e6b969_1
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
cryptography=38.0.3=py39he2a39a8_0
idna=3.4=pyhd8ed1ab_0
isodate=0.6.1=pyhd8ed1ab_0
jmespath=1.0.1=pyhd8ed1ab_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.39.4=h76d750c_0
libzlib=1.2.13=h03a7124_4
msal=2023.03.01=py_0
msal-extensions=2023.03.01=py_0
msrest=2022.06.01=py_0
ncurses=6.3=h07bb92c_1
netaddr=0.7.19=py_0
oauthlib=3.2.2=pyhd8ed1ab_0
openssl=3.1.0=h03a7124_0
pip=22.3.1=pyhd8ed1ab_0
portalocker=2.7.0=py39h2804cbe_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.13=h96fcbfb_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.0=pyhd8ed1ab_0
python_abi=3.9=2_cp39
readline=8.1.2=h46ed386_0
requests=2.28.2=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
s3transfer=0.6.0=pyhd8ed1ab_0
setuptools=65.5.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.39.4=h2229b38_0
tk=8.6.12=he1e0b03_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2022f=h191b570_0
urllib3=1.26.11=pyhd8ed1ab_0
wheel=0.38.4=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0

boto3~=1.26.9
netaddr~=0.7.19
botocore~=1.29.10
attrs~=22.1.0
python-dotenv~=1.0.0
readline~=8.1.2